Modificando funciones para poder generar el Delete de usuarios

+++ flaskProject/model/model_usuarios.py
@@ -1,4 +1,5 @@
 from alchemyClasses.usuarios import usuarios
+from alchemyClasses.rentar import rentar
 from alchemyClasses import db
 from Cryptodome.Hash import SHA256
 from base64 import b64encode
@@ -20,9 +21,7 @@ def crear_usuario(nombre, apPat, password, apMat=None, email=None, profilePictur
 def leer_usuarios():
     return usuarios.query.all()
 
-# Función para obtener un usuario por su ID
-def leer_usuario_por_id(id):
-    return usuarios.query.get(id)
+
 
 # Función para actualizar un usuario por su ID
 def actualizar_usuario(id, nombre=None, apPat=None, apMat=None, password=None, email=None, profilePicture=None, superUser=None):
@@ -47,13 +46,33 @@ def actualizar_usuario(id, nombre=None, apPat=None, apMat=None, password=None, e
     except:
         return -1
 
-# Función para eliminar un usuario por su ID
-def eliminar_usuario(id):
+
+# Función para obtener un usuario por su ID
+def leer_usuario_por_id(id):
+    return usuarios.query.get(id)
+
+
+
+def eliminar_usuario(id_usuario):
+    # Buscar el usuario por su ID
+    usuario = usuarios.query.get(id_usuario)
+
+    # Verificar si el usuario existe
+    if usuario is None:
+        # Si el usuario no existe, devolver -1
+        return -1
     
-    usuario = usuarios.query.get(id)
     try:
-       db.session.delete(usuario)
-       db.session.commit()
-       return 0  # Se encontró y eliminó el usuario correctamente
-    except:
-       return -1  # No se encontró el usuario
+        # Eliminar todas las rentas asociadas al usuario
+        rentar.query.filter_by(idUsuario=id_usuario).delete()
+
+        # Eliminar el usuario
+        db.session.delete(usuario)
+        db.session.commit()
+        
+        # Devolver 0 para indicar éxito
+        return 0
+    except Exception as e:
+        # En caso de error, hacer rollback y devolver el error
+        db.session.rollback()
+        return str(e)
